KernelSU: Introduce KernelSU
The main feature of KernelSU is it is Kernel-based. KernelSU works in kernel mode, so it can provide kernel interface we never had before. For example, we can add hardware breakpoint to any process in kernel mode; We can access physical memory of any process without anybody being aware of; We can intercept any syscall in kernel space; etc.

Signed-off-by: Muhammed Ali Simsek <malisimsek17@gmail.com>